Fixes bug with remove. Now it appears the remove algorithm is numerically stable.
In welford.remove, I was using self.__old_m as the previous m, but that caused a bug.
So instead, I manually store the previous m as old_m = np.copy(self.__m). This fixes the bug
and now it appears the algorithm is numerically stable. Running test_welford.test_remove with a sample
size of 1,000,000 resulted in no assertion errors, thus I believe the modified algorithm is numerically
stable. Again, I have done no formal analysis on this, but imperically it seems to work.